Add files via upload
This is the last commit before the gradio update, which for me has caused a lot of persistent issues.  Changes include, an edit to the one_click file so textgen doesn't update to the latest repo configuration.  An ai character for lucid-vision, an update to the UI for easier access to the mic record button when using whisper, some updated files from oobabooga's textgen repo that were updated after gradio but don't need the updated gradio to function.